Move game functions to game module
