+++ examples/battleship/server/baseline/battleship.js
@@ -1,85 +1,55 @@
 const express = require('express')
 const router = express.Router()
 
-const { getIO, getSocket } = require('./utils/socket')
 const { hash } = require('./utils/hash')
 
 const { fullProve, getVerifyProofInputs } = require('./privacy/proof-verify')
 const truffle_connect = require('../connection/truffle_connect')
 
-let games = new Map()
-
 const { targetEventType, proofEventType, gameEventType } = require('./messaging/eventType')
 const KafkaProducer = require('./messaging/producer');
 
-const userInGame = (id, game) => {
-    return games.has(game) && games.get(game).players.map(player => player.id).includes(id)
-}
+const { games, userInGame, updateGame } = require('./game')
 
-const joinGame = (session, game) => {
-  const socket = getSocket(session)
-
-  socket.join(game)
-
-  socket.on('game:move', (workgroup, message) => {
-    console.log('Received:', msg)
-  })
-}
-
-const startGame = (workgroup) => {
-    console.log(workgroup)
-
-    let game = {
-        id: workgroup.id,
-        shieldContractAddress: workgroup.shieldContractAddress,
-        players: workgroup.players.map(id => ({id})),
-        actions: []
-    }
-
-
-    console.log(`starting game with ID #${workgroup.id}`)
-    updateGame(game)
-}
 
 router.get('/:id', (req, res) => {
-    if (games.has(req.params.id)) {
-        let game = games.get(req.params.id)
-        
-        return res.json({game})
-    }
+  if (games.has(req.params.id)) {
+    let game = games.get(req.params.id)
 
-    res.sendStatus(404)
+    return res.json({ game })
+  }
+
+  res.sendStatus(404)
 })
 
-
 router.put('/hash/:id', async (req, res) => {
-    const id = req.body.id
-    const gameID = req.params.id
+  const id = req.body.id
+  const gameID = req.params.id
 
-    if (games.has(gameID)) {
-        if (userInGame(id, gameID)) {
-            let game = games.get(gameID)
+  if (games.has(gameID)) {
+    if (userInGame(id, gameID)) {
+      let game = games.get(gameID)
 
-            const player = game.players.find(player => player.id === id)
-            
-            if (player.hash !== undefined && player.hash > 0)
-                return res.status(409).send('Player hash already set.')
+      const player = game.players.find(player => player.id === id)
 
-            const boardHash = await hash(req.body);
+      if (player.hash !== undefined && player.hash > 0)
+        return res.status(409).send('Player hash already set.')
 
-            player.hash = boardHash
+      const boardHash = await hash(req.body);
 
-            updateGame(game)
+      player.hash = boardHash
 
-            const gameProducer = new KafkaProducer('game', gameEventType);
-            await gameProducer.queue({id: game.id, shieldContractAddress: game.shieldContractAddress, players: game.players});
-            return res.sendStatus(200)
-        }
+      updateGame(game)
 
-        return res.status(403).send('Action not permitted.')
+      const gameProducer = new KafkaProducer('game', gameEventType);
+      await gameProducer.queue({ id: game.id, shieldContractAddress: game.shieldContractAddress, players: game.players });
+      return res.sendStatus(200)
     }
 
-    res.status(404).send(`Game #${game} does not exist`)
+    return res.status(403).send('Action not permitted.')
+  }
+
+  res.status(404).send(`Game #${game} does not exist`)
 })
 
 router.post('/target', async (req, res) => {
@@ -104,51 +74,29 @@ router.post('/proof', async (req, res) => {
 
   if (parseInt(publicSignals[0]) === req.body.result) {
     const proofMsg = {
-        proof,
-        publicSignals,
-        gameId: req.body.gameId,
-        playerId: req.body.playerId
-      }
-      const proofProducer = new KafkaProducer('proof', proofEventType); 
-      await proofProducer.queue(proofMsg);
-      res.sendStatus(200);
-    
-      return
+      proof,
+      publicSignals,
+      gameId: req.body.gameId,
+      playerId: req.body.playerId
+    }
+    const proofProducer = new KafkaProducer('proof', proofEventType);
+    await proofProducer.queue(proofMsg);
+    res.sendStatus(200);
+
+    return
   }
-  
+
   res.sendStatus(409)
 })
 
-router.post('/verify', async(req, res) => {
-    console.log('verify body', req.body.proof)
-    verifyInputs = await getVerifyProofInputs(req.body.proof, req.body.publicSignals);
-    console.log(games.get(req.body.gameId))
-    const shieldAddress = games.get(req.body.gameId).shieldContractAddress
-    await truffle_connect.verify(verifyInputs.a, verifyInputs.b, verifyInputs.c, verifyInputs.public, shieldAddress, () => {
-        res.send('verified');
-    });
+router.post('/verify', async (req, res) => {
+  verifyInputs = await getVerifyProofInputs(req.body.proof, req.body.publicSignals);
+  const shieldAddress = games.get(req.body.gameId).shieldContractAddress
+  await truffle_connect.verify(verifyInputs.a, verifyInputs.b, verifyInputs.c, verifyInputs.public, shieldAddress, () => {
+    res.send('verified');
+  });
 })
 
-const updateGame = (game) => {
-    const gameExisted = games.has(game.id)
-
-    games.set(game.id, game)
-
-    getIO().to(game.id).emit(gameExisted ? 'game:update' : 'game:init', game)
-}
-
-const handleGameEvent = (type, event) => {
-    if (event.gameId === undefined) {
-        console.error(`Game event ${event} does not specify a gameId.`)
-    }
-
-    getIO().to(event.gameId).emit('game:event', {type, data: event})
-}
-
 module.exports = {
-    battleshipRouter: router,
-    joinGame,
-    startGame,
-    updateGame,
-    handleGameEvent
+  battleshipRouter: router,
 }
